# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ALSS-LEPC Adaptive Learning API",
    "description": "Adaptive Learning System with SAINT KT + D3QN RL",
    "version": "1.0.0"
  },
  "paths": {
    "/api/session/start": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Start Session",
        "operationId": "start_session_api_session_start_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/next-concept": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Next Concept",
        "operationId": "next_concept_api_session__session_id__next_concept_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextConceptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/theory": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "Theory",
        "operationId": "theory_api_session__session_id__theory_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TheoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/exercise": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Generate Exercise",
        "operationId": "generate_exercise_api_session__session_id__exercise_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/submit": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Submit Answer",
        "operationId": "submit_answer_api_session__session_id__submit_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/chat": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Chat About Exercise",
        "operationId": "chat_about_exercise_api_session__session_id__chat_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/status": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "Session Status",
        "operationId": "session_status_api_session__session_id__status_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExerciseResponse": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "title": "Exercise Id"
          },
          "concept_name": {
            "type": "string",
            "title": "Concept Name"
          },
          "concept_idx": {
            "type": "integer",
            "title": "Concept Idx"
          },
          "bloom_level": {
            "type": "integer",
            "title": "Bloom Level"
          },
          "bloom_label": {
            "type": "string",
            "title": "Bloom Label"
          },
          "exercise_type": {
            "$ref": "#/components/schemas/ExerciseType",
            "default": "mcq"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "sentence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentence"
          },
          "options": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Options"
          },
          "statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          },
          "items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Items"
          },
          "pairs": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Pairs"
          },
          "right_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Right Items"
          },
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "max_steps": {
            "type": "integer",
            "title": "Max Steps"
          },
          "theory": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theory"
          },
          "recommendation_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationReason"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "concept_name",
          "concept_idx",
          "bloom_level",
          "bloom_label",
          "question",
          "step",
          "max_steps"
        ],
        "title": "ExerciseResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NextConceptResponse": {
        "properties": {
          "concept_name": {
            "type": "string",
            "title": "Concept Name"
          },
          "concept_idx": {
            "type": "integer",
            "title": "Concept Idx"
          },
          "bloom_level": {
            "type": "integer",
            "title": "Bloom Level"
          },
          "bloom_label": {
            "type": "string",
            "title": "Bloom Label"
          },
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "max_steps": {
            "type": "integer",
            "title": "Max Steps"
          }
        },
        "type": "object",
        "required": [
          "concept_name",
          "concept_idx",
          "bloom_level",
          "bloom_label",
          "step",
          "max_steps"
        ],
        "title": "NextConceptResponse"
      },
...
      "SessionCreateRequest": {
        "properties": {
          "max_steps": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 5.0,
            "title": "Max Steps",
            "default": 9999
          },
          "use_default_data": {
            "type": "boolean",
            "title": "Use Default Data",
            "default": true
          }
        },
        "type": "object",
        "title": "SessionCreateRequest"
      },
      "SessionCreateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "n_concepts": {
            "type": "integer",
            "title": "N Concepts"
          },
          "concepts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Concepts"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "n_concepts",
          "concepts"
        ],
        "title": "SessionCreateResponse"
      },
...
      "SubmitAnswerRequest": {
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/SubmitAnswerPayload"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "SubmitAnswerRequest"
      },
      "SubmitAnswerResponse": {
        "properties": {
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "correct_option": {
            "type": "string",
            "title": "Correct Option"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "concept_name": {
            "type": "string",
            "title": "Concept Name"
          },
          "bloom_level": {
            "type": "integer",
            "title": "Bloom Level"
          },
          "mastery_after": {
            "type": "number",
            "title": "Mastery After"
          },
          "avg_mastery": {
            "type": "number",
            "title": "Avg Mastery"
          },
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "session_completed": {
            "type": "boolean",
            "title": "Session Completed"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          }
        },
        "type": "object",
        "required": [
          "is_correct",
          "correct_option",
          "explanation",
          "concept_name",
          "bloom_level",
          "mastery_after",
          "avg_mastery",
          "step",
          "session_completed",
          "stats"
        ],
        "title": "SubmitAnswerResponse"
      },
      "TheoryResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "examples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Examples"
          }
        },
        "type": "object",
        "required": [
          "content",
          "examples"
        ],
        "title": "TheoryResponse"
      },
...
      "TutorChatResponse": {
        "properties": {
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "explanation"
        ],
        "title": "TutorChatResponse"
      },
...
    },
    "securitySchemes": {
      "XUserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Authenticated user id forwarded by the frontend proxy."
      },
      "XServiceTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-service-token",
        "description": "Shared internal token used by the frontend proxy when calling the backend API."
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment behind the current API host."
    }
  ],
  "security": [
    {
      "XUserIdHeader": [],
      "XServiceTokenHeader": []
    }
  ]
}